Converting backend to typescript
