+++ backend/package.json
@@ -4,17 +4,15 @@
   "main": "index.js",
   "license": "ISC",
   "scripts": {
-    "start": "nodemon index.js"
+    "start": "ts-node index.ts"
   },
   "author": "",
   "dependencies": {
     "binutils": "^0.1.0",
-    "express": "^4.16.4",
-    "lodash": "^4.17.11",
-    "socket.io": "^2.2.0",
     "utf8-bytes": "^0.0.1"
   },
   "devDependencies": {
-    "nodemon": "^1.18.7"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
   }
 }
